requirements.txt
numpy==1.16.1
opencv-contrib-python==4.2.0.34
opencv-python==4.9.0.80
openpyxl==3.0.3
pandas==1.0.3
Pillow==7.1.1
pyttsx3==2.71


